+++ woodwork/tests/data_table/test_datatable.py
@@ -1755,22 +1755,6 @@ def test_natural_language_inference_with_config_options():
     ww.config.reset_option('natural_language_threshold')
 
 
-def test_to_dataframe_copy(sample_df):
-    dt = DataTable(sample_df)
-
-    df_no_copy = dt.to_dataframe()
-    df_copy = dt.to_dataframe(copy=True)
-
-    assert df_no_copy is sample_df
-
-    assert df_no_copy is not df_copy
-
-    df_copy['test_col'] = pd.Series([1, 2, 3])
-    assert 'test_col' in df_copy.columns
-    assert 'test_col' not in df_no_copy.columns
-    assert 'test_col' not in dt.columns
-
-
 def test_describe_does_not_include_index(describe_df):
     dt = DataTable(describe_df, index='index_col')
     stats_df = dt.describe()
@@ -2035,6 +2019,31 @@ def test_data_table_describe_with_include(sample_df):
     multi_params_df['full_name'].equals(dt.describe()['full_name'])
 
 
+def test_value_counts(categorical_pandas_dd_list):
+    for input_dt in categorical_pandas_dd_list:
+        val_cts = input_dt.value_counts()
+        for col in input_dt.columns:
+            if col in ['ints', 'bools']:
+                assert col not in val_cts
+            else:
+                assert col in val_cts
+        assert val_cts['categories1'] == [{'value': 200, 'count': 4}, {'value': 100, 'count': 2}, {'value': 1, 'count': 2}, {'value': 3, 'count': 1}]
+        assert val_cts['categories2'] == [{'value': np.nan, 'count': 5}, {'value': 'test2', 'count': 2}, {'value': 'test', 'count': 2}]
+        assert val_cts['categories3'] == [{'value': np.nan, 'count': 6}, {'value': 'test', 'count': 3}]
+
+        val_cts_descending = input_dt.value_counts(ascending=True)
+        for col, vals in val_cts_descending.items():
+            for i in range(len(vals)):
+                assert vals[i]['count'] == val_cts[col][-i - 1]['count']
+
+        val_cts_dropna = input_dt.value_counts(dropna=True)
+        assert val_cts_dropna['categories3'] == [{'value': 'test', 'count': 3}]
+
+        val_cts_2 = input_dt.value_counts(top_n=2)
+        for col in val_cts_2:
+            assert len(val_cts_2[col]) == 2
+
+
 def test_data_table_handle_nans_for_mutual_info():
     df_nans = pd.DataFrame({
         'nans': pd.Series([None, None, None, None]),
@@ -2046,7 +2055,7 @@ def test_data_table_handle_nans_for_mutual_info():
         'str_no_nan': pd.Series(['test', 'test2', 'test2', 'test']),
     })
     dt_nans = DataTable(df_nans)
-    formatted_df = dt_nans._handle_nans_for_mutual_info(dt_nans.to_dataframe(copy=True))
+    formatted_df = dt_nans._handle_nans_for_mutual_info(dt_nans.to_dataframe().copy())
 
     assert isinstance(formatted_df, pd.DataFrame)
 
@@ -2067,7 +2076,7 @@ def test_data_table_make_categorical_for_mutual_info():
         'categories': pd.Series(['test', 'test2', 'test2', 'test'])
     })
     dt = DataTable(df)
-    formatted_num_bins_df = dt._make_categorical_for_mutual_info(dt.to_dataframe(copy=True), num_bins=4)
+    formatted_num_bins_df = dt._make_categorical_for_mutual_info(dt.to_dataframe().copy(), num_bins=4)
 
     assert isinstance(formatted_num_bins_df, pd.DataFrame)
 
@@ -2090,7 +2099,7 @@ def test_data_table_get_mutual_information(df_same_mi, df_mi):
     assert mi_between_cols('floats', 'ints', mi) == 1.0
 
     dt = DataTable(df_mi)
-    original_df = dt.to_dataframe(copy=True)
+    original_df = dt.to_dataframe().copy()
     mi = dt.get_mutual_information()
     assert mi.shape[0] == 6
     np.testing.assert_almost_equal(mi_between_cols('ints', 'bools', mi), 0.734, 3)
@@ -2183,6 +2192,10 @@ def test_numeric_index_strings(time_index_df):
     error_msg = 'Time index column must contain datetime or numeric values'
     with pytest.raises(TypeError, match=error_msg):
         DataTable(time_index_df, time_index='ints', logical_types={'ints': 'Categorical'})
+
+    error_msg = 'Time index column must contain datetime or numeric values'
+    with pytest.raises(TypeError, match=error_msg):
+        DataTable(time_index_df, time_index='letters', logical_types={'strs': 'Integer'})
 
     dt = DataTable(time_index_df, time_index='strs', logical_types={'strs': 'Double'})
     date_col = dt['strs']