+++ docs/source/release_notes.rst
@@ -6,7 +6,10 @@ Release Notes
 
     * Enhancements
         * Add ``__eq__`` to DataTable and DataColumn and update LogicalType equality (:pr:`318`)
+        * Add ``value_counts()`` method to DataTable (:pr:`342`)
+        * Support serialization and deserialization of DataTables via csv, pickle, or parquet (:pr:`293`)
     * Fixes
+        * Catch non numeric time index at validation (:pr:`332`)
     * Changes
         * Support logical type inference from a Dask DataFrame (:pr:`248`)
         * Fix validation checks and ``make_index`` to work with Dask DataFrames (:pr:`260`)
@@ -18,14 +21,19 @@ Release Notes
         * Rename ``DataColumn.to_pandas`` method to ``DataColumn.to_series`` (:pr:`311`)
         * Rename ``DataTable.to_pandas`` method to ``DataTable.to_dataframe`` (:pr:`319`)
         * Remove UserWarning when no matching columns found (:pr:`325`)
+        * Remove ``copy`` parameter from ``DataTable.to_dataframe`` and ``DataColumn.to_series`` (:pr:`338`)
+        * Allow pandas ExtensionArrays as inputs to DataColumn (:pr:`343`)
     * Documentation Changes
         * Create a guide for using Woodwork with Dask (:pr:`304`)
         * Add conda install instructions (:pr:`305`, :pr:`309`)
         * Fix README.md badge with correct link (:pr:`314`)
+        * Simplify issue templates to make them easier to use (:pr:`339`)
+        * Remove extra output cell in Start notebook (:pr:`341`)
     * Testing Changes
+        * Add support for Python 3.9 in CI (:pr:`207`)
         * Parameterize numeric time index tests (:pr:`288`)
         * Add DockerHub credentials to CI testing environment (:pr:`326`)
-        * Add support for Python 3.9 in CI (:pr:`207`)
+        * Fix removing files for serialization test (:pr:`350`)
 
     Thanks to the following people for contributing to this release:
     :user:`ctduffy`, :user:`gsheni`, :user:`tamargrey`, :user:`thehomebrewnerd`
@@ -34,6 +42,7 @@ Release Notes
 
     * The ``DataColumn.to_pandas`` method was renamed to ``DataColumn.to_series``.
     * The ``DataTable.to_pandas`` method was renamed to ``DataTable.to_dataframe``.
+    * ``copy`` is no longer a parameter of ``DataTable.to_dataframe`` or ``DataColumn.to_series``.
 
 **v0.0.4** October 21, 2020
     * Enhancements